add option --no-alloc
Eliminates all uses of alloc::Box and alloc::Vec when --no-alloc is chosen.

The main difficulty with removing Vec is the fact that wasm functions have multi-returns, so when calling indirect we don't know how many returns come back.

This commit solves it by allocating as much space on the stack as needed for the highest possible number of returns when calling indirect.

There are other non-stack-based solutions, especially the tinyvec crate allows static heapless vector functionality without usage of unsafe.